fix: cleanup types
